Run 96 cycles max before calling update_gba
Theory behind this is it increases the frequency of checks for firing DMA/IRQs so it affects timing.  Activision gams like Sims 2, Monster House and Teen Titans all trigger a high number of interrupts per-frame compared to other games.  This clears up visual defects and artefacts present in lr-gpsp, although not sure how or if it will affect other games.