Separate MAX_BLOCK_SIZE / MAX_EXITS for arm/thumb / reduce arm vals
This improves performance on Camelot games (specifically Mario Tennis / Mario Golf) and solves issue #151 in the main lr-gpsp repo.  Theory is that by reducing the block sizes/exits for arm mode that this reduces the number of times the dynarec caches need to be flushed per frame.  Impact of this change needs to be assessed for all games though.