Implement dynamic Translation Gates
So....Translation Gates were an Exophase idea to stop the recompiler going too far down branches/conditions which might rarely or never get actually run.  It's basically a small list of addresses that the dynarec checks to see if it needs to end a block early.  This has the (hopeful) knock-on effect of not marking the memory area as translated in the SMC mirror, so that we avoid flushing the caches if this area is subsequently written to.

The Translation Gates were originally specified per-game in the game_config.txt file (and later incorporated in the emulator at build time via the gba_over.h header file).

This commit aims to identify these whilst the game is running and add them to the TG list.  Few ToDos here to complete this - the biggest one is that I'm assuming all writes are to IWRAM which is wrong, as EWRAM writes can also trigger SMC flushes.  But currently it does work and will save a few flushes in certain games.