[ARM32] Check value before writing to IWRAM/EWRAM
This is a dynarec check to see if the stored value is the same as the value that we're about to write.  If they are the same we a) don't need to actually do the write and b) avoid a full flush (we also avoid the check to see if we need to do a flush!).

This is possible to do in the other dynarecs but only ARM32 at the moment.  It's not a big performance increase but it does save a few full flushes on some games.